fix: 수집 경로 리뷰를 끝내고 미결 목록을 소진한다 (#117)
## 📝 PR 요약

에이전트에서 DB 까지의 수집 경로 리뷰를 끝내고 거기서 나온 결함을 닫았다. 같이 미결 목록의 검증·결정·부채 항목을 전부
처리해 `docs/temp` 를 비웠다.

## 📌 관련 이슈
* Resolves: #

## 🛠️ 작업 내용
- [x] 수집 경로 리뷰 세션 S3~S8 수행, 이슈 대장 6건 전부 마감
- [x] 매퍼 attr 값 정규화 — 계약이 허용하는 숫자 attr 에서 시계열 축이 유실되던 문제
- [x] 자식 시계열 INSERT 가 모델 자연키를 읽어 충돌 대상을 지정
- [x] `server_pressure` bigint 전환 (컬럼 + 시퀀스), `tasks.status` 폭 확장 —
마이그레이션 2건
- [x] 진입점에서 tz-aware datetime 강제
- [x] pyright 를 CI 게이트로 (`src` 범위)
- [x] 학습 자료·미결 문서 소진 후 삭제, wrap-up 체크리스트 압축
- [x] 버전 1.2.2

## 🚀 PR 유형
- [x] 🐛 버그 수정 (Bug Fix)
- [ ] ✨ 새로운 기능 (New Feature)
- [ ] 📝 문서 작업 (Documentation)
- [ ] 💄 UI/UX 작업 (UI/UX)
- [ ] ♻️ 리팩토링 (Refactoring)
- [ ] 🔧 설정/환경 수정 (Configuration)
- [ ] 🧪 테스트 추가/수정 (Testing)

## 변경 사항

조용히 데이터가 사라지던 경로 둘을 닫았다. 계약은 datapoint attr 값으로 문자열과 숫자를 모두 허용하는데 매퍼가
문자열만 전제해, 코어 인덱스나 device 가 숫자로 오면 그 축의 시계열이 통째로 없어졌다. 자식 시계열 INSERT 는 충돌
대상 없이 `on_conflict_do_nothing` 을 불러 그 테이블의 모든 UNIQUE 위반을 함께 삼켰다 — 자연키를
열거하는 대신 모델이 선언한 제약에서 읽고, 없으면 기동에서 거부한다.

스키마 두 건을 고쳤다. `server_pressure.id` 만 int4 SERIAL 이라 시퀀스 상한을 넘기면 그 호스트의
메트릭 적재 전체가 롤백됐다. 컬럼만 bigint 로 올리면 SERIAL 이 만든 시퀀스는 그대로 남아 `ALTER
SEQUENCE` 를 함께 넣었다 — autogenerate 가 시퀀스 타입을 비교하지 않아 `alembic check` 로는 안
잡히는 갭이다. `tasks.status` 는 인바운드가 32자를 받는데 컬럼이 16자라 그 구간이 검증 통과 후 DLQ 로
갔다.

naive datetime 이 진입점을 통과해 `timestamptz` 컬럼까지 흘렀다. asyncpg 가 naive 를 거부하지
않고 로컬 시각으로 해석하므로 컨테이너 TZ 에 따라 조용히 어긋난 값이 쌓인다. 검증은 진입점 한 곳이라는 규약대로 스키마에서
막았다.

pyright `src` 위반이 0 이 되어 CI 게이트로 걸었다. check 이름이 바뀌면 required 재등록이 필요해 기존
lint 잡 안에 넣었다.

## 테스트 방법

1. `uv run pytest tests/unit tests/integration -q` — 976 통과
2. `uv run pyright src` — 0
3. 마이그레이션: `alembic upgrade head` 후 `select seqtypid::regtype from
pg_sequence where seqrelid='server_pressure_id_seq'::regclass` 가 bigint
인지, `alembic downgrade -1` 왕복이 되는지
4. naive datetime 이 거부되는지 — `MetricsInput.model_validate` 에 offset 없는
`collected_at` 을 넣으면 `ValidationError`

## 스크린샷 (선택)

## 체크리스트
- [x] 로컬에서 동작 확인
- [x] .env.example 업데이트 (환경변수 추가 시)
- [ ] README 업데이트 (필요 시)

## 관련 이슈

## 💡 기타 참고 사항 (선택 사항)

`docs/temp` 가 규약 파일 하나만 남았다. 학습 자료 4개는 저장소 고유 사실만 영구 문서에 옮기고 지웠고, 그 과정에서
`deploy.md` 의 rollback 서술과 `docker.md` 의 포트 노출 본문이 코드보다 낡아 있던 것을 함께 고쳤다.
실 환경 캡처가 필요해 확인할 수 없는 계약 검증 항목은 tradeoffs T21 로 남겼다.

버전은 patch 로 올렸다. `rotate-secret.sh` 가 `feat` 커밋으로 들어가 있으나 제품 기능이 아니라 운영
스크립트라는 판단이다.

GitHub 저장소 설정 두 건은 UI 작업이라 코드로 처리할 수 없어 `ci-setup.md` 3.5 절에 남겼다 —
ruleset 등록과 required check 재등록, tag ruleset 의 Actions bypass 확인이다.

+++ src/assessment_engine/db/repositories/query/base_report.py
@@ -19,7 +19,7 @@ class BaseReportQueryRepository(ABC):
     async def report_aggregate(
         self,
         server_ids: list[int],
-        period_days: int,
+        period_days: float,
         end: datetime,
     ) -> list[ReportRowRaw]: ...
 
@@ -27,7 +27,7 @@ class BaseReportQueryRepository(ABC):
     async def report_uptime_stats(
         self,
         server_ids: list[int],
-        period_days: int,
+        period_days: float,
         end: datetime,
     ) -> dict[int, int]:
         """server_id -> period 안 boot_time 변경(재부팅) 횟수."""
@@ -36,7 +36,7 @@ class BaseReportQueryRepository(ABC):
     async def report_agent_restart_stats(
         self,
         server_ids: list[int],
-        period_days: int,
+        period_days: float,
         end: datetime,
     ) -> dict[int, int]:
         """server_id -> period 안 agent_started_at 변경(에이전트 재시작) 횟수.
@@ -52,7 +52,7 @@ class BaseReportQueryRepository(ABC):
     async def report_disk_io_baseline(
         self,
         server_ids: list[int],
-        period_days: int,
+        period_days: float,
         end: datetime,
     ) -> dict[int, DiskIoBaselineRaw]:
         """server_id -> DiskIoBaselineRaw (iops·throughput baseline + p95/peak)."""
@@ -61,7 +61,7 @@ class BaseReportQueryRepository(ABC):
     async def report_net_io_baseline(
         self,
         server_ids: list[int],
-        period_days: int,
+        period_days: float,
         end: datetime,
     ) -> dict[int, NetIoBaselineRaw]:
         """server_id -> NetIoBaselineRaw (rx·tx baseline + p95/peak)."""
